Fetched Data and displayed total
