--- src/pages/api.js
@@ -1,1 +1,0 @@
-
